Moved decentralized centric functions into its own file to make it easier to find them.
Moved common assignments of values outside of the
if/else to make it more readable.

Some minor cleanup of comments and added a comment
to explain the purpose of a function.
